+++ CHANGELOG.rst
@@ -14,6 +14,7 @@ Changelog
 ------
 - Add `Model.raw` method to support the raw sql query.
 - Fix `QuerySet` subclass being lost when `_clone` is run on the instance.
+- Fix bug in `.values` with `source_field`. (#844)
 
 0.17.7
 ------